refactor: move follow logic to a custom hook
